Hints
Added a profile script. This should be added to the character, and will add properties to the character based on the name and mesh. If chosen as the target, ProvideHints gives a list of hints according to the int parameter. To ensure you always get male/female and two hints that will at minimum allow identification, always use at least 3 as the parameter. It would be nice to randomize the other hints, but that's if we have time