Profile update
Profile updated with identifying info for the 19 characters. We should probably not add any more. It should be fine to change the skin tones with the A, B, and C meshes, but the hints don't distinguish based on skin tone, so there shouldn't be B and C NPCs running around if the target has an A mesh.